feat: Added NewFITSHeader() to iris module.
feat: Added NewFITSHeader() to iris module. Includes associated test suite for module export definition and expected output.
