+++ coverage.txt
@@ -1,12 +1,25 @@
 mode: atomic
-github.com/observerly/iris/pkg/fits/fits.go:17.33,28.2 1 2
-github.com/observerly/iris/pkg/fits/fits.go:47.32,52.2 1 1
-github.com/observerly/iris/pkg/fits/fits.go:56.72,59.31 2 1
-github.com/observerly/iris/pkg/fits/fits.go:63.2,63.17 1 1
-github.com/observerly/iris/pkg/fits/fits.go:67.2,78.3 1 1
-github.com/observerly/iris/pkg/fits/fits.go:59.31,61.3 1 2
-github.com/observerly/iris/pkg/fits/fits.go:63.17,65.3 1 1
-github.com/observerly/iris/pkg/fits/fits.go:83.55,98.2 2 1
+github.com/observerly/iris/pkg/histogram/histogram.go:18.51,25.47 4 0
+github.com/observerly/iris/pkg/histogram/histogram.go:30.2,30.12 1 0
+github.com/observerly/iris/pkg/histogram/histogram.go:25.47,28.3 2 0
+github.com/observerly/iris/pkg/fits/fits.go:20.32,25.2 1 1
+github.com/observerly/iris/pkg/fits/fits.go:29.72,32.31 2 1
+github.com/observerly/iris/pkg/fits/fits.go:36.2,36.17 1 1
+github.com/observerly/iris/pkg/fits/fits.go:40.2,51.3 1 1
+github.com/observerly/iris/pkg/fits/fits.go:32.31,34.3 1 2
+github.com/observerly/iris/pkg/fits/fits.go:36.17,38.3 1 1
+github.com/observerly/iris/pkg/fits/fits.go:56.55,71.2 2 1
+github.com/observerly/iris/pkg/fits/header.go:17.33,28.2 1 3
+github.com/observerly/iris/pkg/photometry/stars.go:22.93,27.29 2 2
+github.com/observerly/iris/pkg/photometry/stars.go:31.2,40.3 2 2
+github.com/observerly/iris/pkg/photometry/stars.go:27.29,29.3 1 32
+github.com/observerly/iris/pkg/photometry/stars.go:43.51,46.34 2 1
+github.com/observerly/iris/pkg/photometry/stars.go:66.2,66.14 1 1
+github.com/observerly/iris/pkg/photometry/stars.go:46.34,48.35 1 256
+github.com/observerly/iris/pkg/photometry/stars.go:48.35,52.22 2 24
+github.com/observerly/iris/pkg/photometry/stars.go:62.4,62.29 1 24
+github.com/observerly/iris/pkg/photometry/stars.go:52.22,56.85 2 23
+github.com/observerly/iris/pkg/photometry/stars.go:56.85,58.6 1 5
 github.com/observerly/iris/pkg/photometry/noise.go:12.77,17.29 2 2
 github.com/observerly/iris/pkg/photometry/noise.go:21.2,26.3 1 2
 github.com/observerly/iris/pkg/photometry/noise.go:17.29,19.3 1 32
@@ -18,42 +31,6 @@ github.com/observerly/iris/pkg/photometry/noise.go:58.34,66.29 3 225
 github.com/observerly/iris/pkg/photometry/noise.go:73.4,73.25 1 225
 github.com/observerly/iris/pkg/photometry/noise.go:66.29,67.44 1 2025
 github.com/observerly/iris/pkg/photometry/noise.go:67.44,69.6 1 1978
-github.com/observerly/iris/pkg/photometry/stars.go:22.93,27.29 2 2
-github.com/observerly/iris/pkg/photometry/stars.go:31.2,40.3 2 2
-github.com/observerly/iris/pkg/photometry/stars.go:27.29,29.3 1 32
-github.com/observerly/iris/pkg/photometry/stars.go:43.51,46.34 2 1
-github.com/observerly/iris/pkg/photometry/stars.go:66.2,66.14 1 1
-github.com/observerly/iris/pkg/photometry/stars.go:46.34,48.35 1 256
-github.com/observerly/iris/pkg/photometry/stars.go:48.35,52.22 2 24
-github.com/observerly/iris/pkg/photometry/stars.go:62.4,62.29 1 24
-github.com/observerly/iris/pkg/photometry/stars.go:52.22,56.85 2 23
-github.com/observerly/iris/pkg/photometry/stars.go:56.85,58.6 1 5
-github.com/observerly/iris/pkg/histogram/histogram.go:18.51,25.47 4 0
-github.com/observerly/iris/pkg/histogram/histogram.go:30.2,30.12 1 0
-github.com/observerly/iris/pkg/histogram/histogram.go:25.47,28.3 2 0
-github.com/observerly/iris/pkg/iris/convolution.go:5.122,16.40 5 128
-github.com/observerly/iris/pkg/iris/convolution.go:21.2,21.21 1 128
-github.com/observerly/iris/pkg/iris/convolution.go:25.2,25.21 1 128
-github.com/observerly/iris/pkg/iris/convolution.go:29.2,34.12 5 128
-github.com/observerly/iris/pkg/iris/convolution.go:16.40,19.3 2 98
-github.com/observerly/iris/pkg/iris/convolution.go:21.21,23.3 1 112
-github.com/observerly/iris/pkg/iris/convolution.go:25.21,27.3 1 112
-github.com/observerly/iris/pkg/iris/convolution.go:37.126,54.19 8 128
-github.com/observerly/iris/pkg/iris/convolution.go:58.2,58.19 1 128
-github.com/observerly/iris/pkg/iris/convolution.go:62.2,66.21 3 128
-github.com/observerly/iris/pkg/iris/convolution.go:70.2,70.21 1 128
-github.com/observerly/iris/pkg/iris/convolution.go:74.2,79.14 5 128
-github.com/observerly/iris/pkg/iris/convolution.go:54.19,56.3 1 112
-github.com/observerly/iris/pkg/iris/convolution.go:58.19,60.3 1 112
-github.com/observerly/iris/pkg/iris/convolution.go:66.21,68.3 1 112
-github.com/observerly/iris/pkg/iris/convolution.go:70.21,72.3 1 112
-github.com/observerly/iris/pkg/iris/convolution.go:82.124,93.19 5 128
-github.com/observerly/iris/pkg/iris/convolution.go:97.2,97.19 1 128
-github.com/observerly/iris/pkg/iris/convolution.go:101.2,101.36 1 128
-github.com/observerly/iris/pkg/iris/convolution.go:106.2,111.13 5 128
-github.com/observerly/iris/pkg/iris/convolution.go:93.19,95.3 1 112
-github.com/observerly/iris/pkg/iris/convolution.go:97.19,99.3 1 112
-github.com/observerly/iris/pkg/iris/convolution.go:101.36,104.3 2 98
 github.com/observerly/iris/pkg/iris/monochrome.go:26.84,39.2 3 10
 github.com/observerly/iris/pkg/iris/monochrome.go:41.79,46.16 3 11
 github.com/observerly/iris/pkg/iris/monochrome.go:50.2,50.18 1 11
@@ -108,3 +85,26 @@ github.com/observerly/iris/pkg/iris/rggb.go:132.2,132.22 1 1
 github.com/observerly/iris/pkg/iris/rggb.go:115.16,117.3 1 0
 github.com/observerly/iris/pkg/iris/rggb.go:121.16,123.3 1 0
 github.com/observerly/iris/pkg/iris/rggb.go:128.16,130.3 1 0
+github.com/observerly/iris/pkg/iris/convolution.go:5.122,16.40 5 128
+github.com/observerly/iris/pkg/iris/convolution.go:21.2,21.21 1 128
+github.com/observerly/iris/pkg/iris/convolution.go:25.2,25.21 1 128
+github.com/observerly/iris/pkg/iris/convolution.go:29.2,34.12 5 128
+github.com/observerly/iris/pkg/iris/convolution.go:16.40,19.3 2 98
+github.com/observerly/iris/pkg/iris/convolution.go:21.21,23.3 1 112
+github.com/observerly/iris/pkg/iris/convolution.go:25.21,27.3 1 112
+github.com/observerly/iris/pkg/iris/convolution.go:37.126,54.19 8 128
+github.com/observerly/iris/pkg/iris/convolution.go:58.2,58.19 1 128
+github.com/observerly/iris/pkg/iris/convolution.go:62.2,66.21 3 128
+github.com/observerly/iris/pkg/iris/convolution.go:70.2,70.21 1 128
+github.com/observerly/iris/pkg/iris/convolution.go:74.2,79.14 5 128
+github.com/observerly/iris/pkg/iris/convolution.go:54.19,56.3 1 112
+github.com/observerly/iris/pkg/iris/convolution.go:58.19,60.3 1 112
+github.com/observerly/iris/pkg/iris/convolution.go:66.21,68.3 1 112
+github.com/observerly/iris/pkg/iris/convolution.go:70.21,72.3 1 112
+github.com/observerly/iris/pkg/iris/convolution.go:82.124,93.19 5 128
+github.com/observerly/iris/pkg/iris/convolution.go:97.2,97.19 1 128
+github.com/observerly/iris/pkg/iris/convolution.go:101.2,101.36 1 128
+github.com/observerly/iris/pkg/iris/convolution.go:106.2,111.13 5 128
+github.com/observerly/iris/pkg/iris/convolution.go:93.19,95.3 1 112
+github.com/observerly/iris/pkg/iris/convolution.go:97.19,99.3 1 112
+github.com/observerly/iris/pkg/iris/convolution.go:101.36,104.3 2 98
